+++ spec/dummy_no_webpacker/config/initializers/react_on_rails.rb
@@ -16,7 +16,7 @@ end
 
 ReactOnRails.configure do |config|
   config.node_modules_location = "client" # Pre 9.0.0 always used "client"
-  config.npm_build_production_command = "yarn run build:production"
+  config.build_production_command = "yarn run build:production"
   config.build_test_command = "yarn run build:test"
   config.generated_assets_dir = File.join(%w[app assets webpack])
   config.webpack_generated_files = %w[app-bundle.js vendor-bundle.js server-bundle.js]